* Remove disfunctional old SampleAllocator. * Rename AnotherSampleAllocator to SampleAllocator. * Refactor methods in SampleAllocator.
