+++ src/test/java/org/kritiniyoga/karmayoga/AllocatorChecks.java
@@ -5,7 +5,7 @@ import io.vavr.collection.Seq;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.kritiniyoga.karmayoga.allocators.AnotherSampleAllocator;
+import org.kritiniyoga.karmayoga.allocators.SampleAllocator;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -32,7 +32,7 @@ public class AllocatorChecks {
     }
 
     static Stream<Allocator> generateAllocators() {
-        return Stream.of(new AnotherSampleAllocator());
+        return Stream.of(new SampleAllocator());
     }
 
     @ParameterizedTest
